# ls: Print optimizations