@sylvestre 2025-04-20
> GNU test failed: tests/ls/stat-failed. tests/ls/stat-failed is passing on 'main'. Maybe you have to rebase? 

Seems that it regressed this :)