@drinkcat 2025-04-20
> GNU testsuite comparison:
> 
> ```
> GNU test failed: tests/misc/tee. tests/misc/tee is passing on 'main'. Maybe you have to rebase?
> ```

That... doesn't fail locally. And has nothing to do with `ls` AFAICT...