@sylvestre 2025-04-21
> > > GNU test failed: tests/ls/stat-failed. tests/ls/stat-failed is passing on 'main'. Maybe you have to rebase?
> > 
> > 
> > Seems that it regressed this :)
> 
> Whoopsie ,-) I was perhaps relying too much on Rust unit tests ,-) Fixed.

Maybe add a test to cover it too? :)